pyproject.toml
[project]
name = "finance-assistant"
version = "3.1.0"
description = "Simple AI-powered financial analysis platform with real-time data - perfect for learning Python!"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core web framework
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.7",
    
    # Stock data
    "yfinance>=0.2.65",
    "pandas>=2.3.1",
    "numpy>=1.24.0",
    
    # AI/ML for financial analysis
    "openai>=1.45.0",
    
    # Basic utilities
    "python-dateutil>=2.8.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
